refactor: Fake 관련 객체 패키지 분리
