+++ src/test/java/kitchenpos/application/fake/repository/InMemoryMenuRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.application.menu;
+package kitchenpos.application.fake.repository;
 
 import java.util.ArrayList;
 import java.util.List;
