+++ src/test/java/kitchenpos/application/menu/MenuServiceTest.java
@@ -7,15 +7,14 @@ import static org.junit.jupiter.api.Assertions.assertAll;
 import java.math.BigDecimal;
 import java.util.List;
 import java.util.NoSuchElementException;
-import kitchenpos.application.MenuGroupService;
 import kitchenpos.application.MenuService;
-import kitchenpos.application.ProductService;
+import kitchenpos.application.fake.repository.InMemoryMenuRepository;
 import kitchenpos.application.fixture.MenuFixture;
 import kitchenpos.application.fixture.MenuGroupFixture;
 import kitchenpos.application.fixture.ProductFixture;
-import kitchenpos.application.infra.FakePurgomalumClient;
-import kitchenpos.application.menu_group.InMemoryMenuGroupRepository;
-import kitchenpos.application.product.InMemoryProductRepository;
+import kitchenpos.application.fake.infra.FakePurgomalumClient;
+import kitchenpos.application.fake.repository.InMemoryMenuGroupRepository;
+import kitchenpos.application.fake.repository.InMemoryProductRepository;
 import kitchenpos.domain.Menu;
 import kitchenpos.domain.MenuGroup;
 import kitchenpos.domain.MenuGroupRepository;
