+++ src/test/java/kitchenpos/application/order/OrderServiceTest.java
@@ -7,17 +7,14 @@ import java.util.List;
 import java.util.NoSuchElementException;
 import java.util.UUID;
 import kitchenpos.application.OrderService;
+import kitchenpos.application.fake.repository.InMemoryOrderRepository;
 import kitchenpos.application.fixture.MenuFixture;
-import kitchenpos.application.fixture.MenuGroupFixture;
 import kitchenpos.application.fixture.OrderFixture;
 import kitchenpos.application.fixture.OrderTableFixture;
-import kitchenpos.application.fixture.ProductFixture;
-import kitchenpos.application.infra.FakeKitchenridersClient;
-import kitchenpos.application.menu.InMemoryMenuRepository;
-import kitchenpos.application.order_table.InMemoryOrderTableRepository;
+import kitchenpos.application.fake.infra.FakeKitchenridersClient;
+import kitchenpos.application.fake.repository.InMemoryMenuRepository;
+import kitchenpos.application.fake.repository.InMemoryOrderTableRepository;
 import kitchenpos.domain.Menu;
-import kitchenpos.domain.MenuGroup;
-import kitchenpos.domain.MenuProduct;
 import kitchenpos.domain.MenuRepository;
 import kitchenpos.domain.Order;
 import kitchenpos.domain.OrderLineItem;
@@ -26,20 +23,15 @@ import kitchenpos.domain.OrderStatus;
 import kitchenpos.domain.OrderTable;
 import kitchenpos.domain.OrderTableRepository;
 import kitchenpos.domain.OrderType;
-import kitchenpos.domain.Product;
 import kitchenpos.infra.KitchenridersClient;
-import org.jetbrains.annotations.NotNull;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.EnumSource;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
 import org.junit.jupiter.params.provider.NullSource;
-import org.mockito.Mock;
-import org.mockito.junit.jupiter.MockitoExtension;
 
 public class OrderServiceTest {
