+++ src/test/java/kitchenpos/application/product/ProductServiceTest.java
@@ -9,12 +9,13 @@ import java.util.UUID;
 import kitchenpos.application.MenuGroupService;
 import kitchenpos.application.MenuService;
 import kitchenpos.application.ProductService;
+import kitchenpos.application.fake.repository.InMemoryProductRepository;
 import kitchenpos.application.fixture.MenuFixture;
 import kitchenpos.application.fixture.MenuGroupFixture;
 import kitchenpos.application.fixture.ProductFixture;
-import kitchenpos.application.infra.FakePurgomalumClient;
-import kitchenpos.application.menu.InMemoryMenuRepository;
-import kitchenpos.application.menu_group.InMemoryMenuGroupRepository;
+import kitchenpos.application.fake.infra.FakePurgomalumClient;
+import kitchenpos.application.fake.repository.InMemoryMenuRepository;
+import kitchenpos.application.fake.repository.InMemoryMenuGroupRepository;
 import kitchenpos.domain.Menu;
 import kitchenpos.domain.MenuGroup;
 import kitchenpos.domain.MenuGroupRepository;
